move MiddlewareArray class to `utils.ts`
