+++ src/getDefaultMiddleware.ts
@@ -11,6 +11,7 @@ import {
   createSerializableStateInvariantMiddleware,
   SerializableStateInvariantMiddlewareOptions
 } from './serializableStateInvariantMiddleware'
+import { MiddlewareArray } from './utils'
 
 function isBoolean(x: any): x is boolean {
   return typeof x === 'boolean'
@@ -54,37 +55,6 @@ export function curryGetDefaultMiddleware<
 >(): CurriedGetDefaultMiddleware<S> {
   return function curriedGetDefaultMiddleware(options) {
     return getDefaultMiddleware(options)
-  }
-}
-
-export class MiddlewareArray<
-  Middlewares extends Middleware<any, any>
-> extends Array<Middlewares> {
-  concat<AdditionalMiddlewares extends ReadonlyArray<Middleware<any, any>>>(
-    items: AdditionalMiddlewares
-  ): MiddlewareArray<Middlewares | AdditionalMiddlewares[number]>
-
-  concat<AdditionalMiddlewares extends ReadonlyArray<Middleware<any, any>>>(
-    ...items: AdditionalMiddlewares
-  ): MiddlewareArray<Middlewares | AdditionalMiddlewares[number]>
-
-  concat(...arr: any[]) {
-    return new MiddlewareArray(...super.concat(...arr))
-  }
-
-  prepend<AdditionalMiddlewares extends ReadonlyArray<Middleware<any, any>>>(
-    items: AdditionalMiddlewares
-  ): MiddlewareArray<AdditionalMiddlewares[number] | Middlewares>
-
-  prepend<AdditionalMiddlewares extends ReadonlyArray<Middleware<any, any>>>(
-    ...items: AdditionalMiddlewares
-  ): MiddlewareArray<AdditionalMiddlewares[number] | Middlewares>
-
-  prepend(...arr: any[]) {
-    if (arr.length === 1 && Array.isArray(arr[0])) {
-      return new MiddlewareArray(...arr[0].concat(this))
-    }
-    return new MiddlewareArray(...arr.concat(this))
   }
 }
